previous-work/privacy_agent_jimmy/requirements.txt
pyautogui==0.9.54
pillow==10.0.1
google-generativeai==0.3.2
psutil==5.9.6
subprocess32==3.5.4

